TS renaming spree
