+++ typescript/apps/api-hono/.http/api-projects.http
@@ -2,14 +2,14 @@
 @urlProjects = http://localhost:5002/api/v1/projects
 @contentType = application/json
 
-### --------------------------------------------------------------------------------- ###
-# GET all projects
+# ------------------------------------------------------------------------------------- #
+### GET all projects
 GET {{urlProjects}}
 Content-Type: {{contentType}}
 
 
-### --------------------------------------------------------------------------------- ###
-# GET project by id
+# ------------------------------------------------------------------------------------- #
+### GET project by id
 @id = 52e1b54a
 
 # GET {{urlProjects}}/{{id}}
@@ -17,8 +17,8 @@ GET {{urlProjects}}/{{project.response.body.$.id}}
 Content-Type: {{contentType}}
 
 
-### --------------------------------------------------------------------------------- ###
-# CREATE project
+# ------------------------------------------------------------------------------------- #
+### CREATE project
 # @name project
 POST {{urlProjects}}
 Content-Type: {{contentType}}
@@ -31,8 +31,8 @@ Content-Type: {{contentType}}
 }
 
 
-### --------------------------------------------------------------------------------- ###
-# CREATE project w/ error
+# ------------------------------------------------------------------------------------- #
+### CREATE project w/ error
 # @name project
 POST {{urlProjects}}/error
 Content-Type: {{contentType}}
@@ -45,8 +45,8 @@ Content-Type: {{contentType}}
 }
 
 
-### --------------------------------------------------------------------------------- ###
-# PUT project by id
+# ------------------------------------------------------------------------------------- #
+### PUT project by id
 # PUT {{urlProjects}}/{{id}}
 PUT {{urlProjects}}/{{project.response.body.$.id}}
 Content-Type: {{contentType}}
@@ -62,8 +62,8 @@ Content-Type: {{contentType}}
 }
 
 
-### --------------------------------------------------------------------------------- ###
-# DELETE project by Id
+# ------------------------------------------------------------------------------------- #
+### DELETE project by Id
 # DELETE {{urlProjects}}/{{id}}
 DELETE {{urlProjects}}/{{project.response.body.$.id}}
 Content-Type: {{contentType}}
