+++ typescript/apps/api-hono/src/app.ts
@@ -1,13 +1,13 @@
 import { createApp } from "#lib/init.ts";
-import { demosRouter } from "#routes/demos.ts";
-import { projectsRouter } from "#routes/projects-rpc.ts";
+import { demoRouter } from "#routes/demo.ts";
+import { projectRouter } from "#routes/project-rpc.ts";
 
 const app = createApp();
 
 // export type App = typeof appRouter;
 // const appRouter = app.route("/", demosRouter).route("/", projectsRouter);
 
-const appRouters = [demosRouter, projectsRouter] as const;
+const appRouters = [demoRouter, projectRouter] as const;
 appRouters.forEach((r) => {
 	app.route("/", r);
 });
