+++ typescript/packages/core/src/project/utils.ts
